懒得喷

+++ apps/help.js
@@ -22,30 +22,22 @@ export class Help extends plugin {
     }
 
     async help(e) {
-        let helpGroup = [];
-        _.forEach(helpList, (group) => {
-            _.forEach(group.list, (help) => {
-                let icon = help.icon * 1;
-                if (!icon) {
-                    help.css = 'display:none';
-                } else {
-                    let x = (icon - 1) % 10;
-                    let y = (icon - x - 1) / 10;
-                    help.css = `background-position:-${x * 50}px -${y * 50}px`;
-                }
-            });
-            helpGroup.push(group);
-        });
-
         try {
-            const render = new Render();
-            const imagePath = await render.help(helpCfg, helpGroup);
-            await e.reply(segment.image(imagePath));
-            fs.unlinkSync(imagePath);
+            const render = new Render()
+            const imagePath = await render.help(helpCfg, helpList)
+            
+            // 发送图片
+            if(fs.existsSync(imagePath)) {
+                await e.reply(segment.image(imagePath))
+                // 发送后删除临时文件
+                fs.unlinkSync(imagePath)
+            } else {
+                throw new Error('生成帮助图片失败')
+            }
         } catch(err) {
-            logger.error(err);
-            await e.reply('生成帮助图片失败');
+            logger.error(err)
+            await e.reply('生成帮助图片失败,请稍后重试')
         }
-        return true;
+        return true
     }
 }
